# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Keycloak Middleware",
        "description": "Documentación Swagger de la API intermedia entre Laravel y Keycloak",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clients/roles": {
            "get": {
                "tags": [
                    "Client Roles"
                ],
                "summary": "Obtiene una lista de los roles por Id de cliente de un reino",
                "operationId": "9fe6db12d281848bc7aa469da63b25be",
                "parameters": [
                    {
                        "name": "realm",
                        "in": "query",
                        "description": "Id del reino",
                        "required": true
                    },
                    {
                        "name": "client_uuid",
                        "in": "query",
                        "description": "Id del client",
                        "required": true
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Tamaño del paginado",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numero de Pagina.",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/clients/roles/role-by-name": {
            "get": {
                "tags": [
                    "Client Roles"
                ],
                "summary": "Obtiene un roles por nombre.",
                "operationId": "d7886681f41d6a5c38e5c2d3c9e28614",
                "parameters": [
                    {
                        "name": "realm",
                        "in": "query",
                        "description": "Id del reino",
                        "required": true
                    },
                    {
                        "name": "client_uuid",
                        "in": "query",
                        "description": "Id del client",
                        "required": true
                    },
                    {
                        "name": "role_name",
                        "in": "query",
                        "description": "Nombre del client",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtiene el role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/clients/roles/create": {
            "post": {
                "tags": [
                    "Client Roles"
                ],
                "summary": "Crear un nuevo rol de cliente.",
                "operationId": "8c93b6451672a9e564b3a4923305b9a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "realm",
                                    "client_uuid",
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "realm": {
                                        "type": "string",
                                        "example": "mi-reino"
                                    },
                                    "client_uuid": {
                                        "type": "string",
                                        "example": "1c742755-b30d-44a4-8b4b-4a60d6059a27"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "test-client-role"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Role de prueba"
                                    },
                                    "attributes": {
                                        "properties": {
                                            "permission": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "read"
                                                ]
                                            },
                                            "permission_2": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "write"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rol creado exitosamente."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/clients/roles/update": {
            "put": {
                "tags": [
                    "Client Roles"
                ],
                "summary": "Update un rol de cliente.",
                "operationId": "971d1dea6c6ffd1b5d08d56990f82801",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "realm",
                                    "client_uuid",
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "realm": {
                                        "type": "string",
                                        "example": "mi-reino"
                                    },
                                    "client_uuid": {
                                        "type": "string",
                                        "example": "1c742755-b30d-44a4-8b4b-4a60d6059a27"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "test-client-role"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Role de prueba"
                                    },
                                    "attributes": {
                                        "properties": {
                                            "permission": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "read"
                                                ]
                                            },
                                            "permission_2": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "write"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rol creado exitosamente."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/clients/roles/delete": {
            "delete": {
                "tags": [
                    "Client Roles"
                ],
                "summary": "Eliminar rol del cliente.",
                "description": "Endpoint para Eliminar un rol cliente.",
                "operationId": "66383ef3c3218d019f8e2f1d02bf9bd7",
                "parameters": [
                    {
                        "name": "realm",
                        "in": "query",
                        "description": "Id del reino",
                        "required": true
                    },
                    {
                        "name": "client_uuid",
                        "in": "query",
                        "description": "Id del usuario",
                        "required": true
                    },
                    {
                        "name": "role_name",
                        "in": "query",
                        "description": "Id del usuario",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Rol eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "123e4567-e89b-12d3-a456-426614174000"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "juanperez"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "juan.perez@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en los datos enviados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Datos inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error interno"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer Token)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}